+++ NEWS.md
@@ -16,7 +16,7 @@
 * `counts.data.frame()` and `props.data.frame()` retain `attributes()` of selected columns
 * `props()` and family gain argument `na.rm` to not count `NA` values for proportions
 * adds `make_sf()` as a simple wrapper for package specific `system.file()`
-* `todos()` and `fixmes()` will not search for `.R` or `.Rmd` files if the `path` is a directory without an `.Rproj`
+* `todos()` and `fixmes()` will not search for `.R` or `.Rmd` files if the `path` is not changed from `""` and no `.Rproj` is found within the directory
 
 # mark 0.2.0